Install FastImage; modify importer and add tests to select highest quality cover
