+++ app/assets/stylesheets/application.css
@@ -67,8 +67,8 @@
 }
 
 .book-cover img {
-  max-height: 475px;
   max-width: 300px;
+  max-height: 450px;
   border: 1px solid #333333;
 }
