+++ app/services/book_importer_service/isbn_constants.rb
@@ -1,0 +1,311 @@
+class BookImporterService
+  class IsbnConstants
+  ISBNS_TO_IMPORT = [
+    '0786222727',
+    '0786222735',
+    '1594130027',
+    '1594130035',
+    '043935806X',
+    '0756967651',
+    '0545139708',
+    '9781338099133',
+    '0261102354',
+    '0261102362',
+    '0261102370',
+    '0544174224',
+    '0099134012',
+    '0440241588',
+    '0099280256',
+    '0099993805',
+    '0385319053',
+    '0099410311',
+    '0375433511',
+    '0385339690',
+    '0099179415',
+    '044022165X',
+    '0375433538',
+    '0375431489',
+    '0440225701',
+    '0375433074',
+    '0440242002',
+    '0452277507',
+    '1846553288',
+    '0756404738',
+    '1596063130',
+    '0451525884',
+    '0192840827',
+    '1593080492',
+    '0140434674',
+    '0345418972',
+    '0345370775',
+    '0007154534',
+    '0345359321',
+    '0606298231',
+    '0099303744',
+    '0061782556',
+    '0345418980',
+    '0752224417',
+    '0060092572',
+    '0061929379',
+    '0061782637',
+    '0061015733',
+    '0060873035',
+    '0099244721',
+    '0099556316',
+    '0451183665',
+    '0060502304',
+    '0099601117',
+    '0670038601',
+    '0670018864',
+    '9780143119494',
+    '014303667X',
+    '0356508366',
+    '0316229245',
+    '0684841258',
+    '0684841215',
+    '1556114249',
+    '0684868199',
+    '0684865130',
+    '0684839741',
+    '9781476746586',
+    '1587247798',
+    '0743212746',
+    '9781439182802',
+    '0061147958',
+    '0062200631',
+    '0061147931',
+    '9780062222992',
+    '9781613777480',
+    '9781600103841',
+    '9781600107610',
+    '9781600106958',
+    '9781600108860',
+    '1613772270',
+    '0385689225',
+    '1501133888',
+    '0812975928',
+    '9780812979299',
+    '0553447432',
+    '0061985848',
+    '9780061985867',
+    '9780061985881',
+    '0062135848',
+    '9788408114178',
+    '1400079179',
+    '0385504225',
+    '1416528652',
+    '0786282290',
+    '0312335164',
+    '9780735211209',
+    '9780857522313',
+    '0553256157',
+    '9780470625767',
+    '0684825554',
+    '0395706734',
+    '0440227712',
+    '043985802X',
+    '0439858038',
+    '0618070265',
+    '0439858054',
+    '9780439783231',
+    '9780439783217',
+    '0833553321',
+    '0395689856',
+    '0395857546',
+    '0688146279',
+    '068817423X',
+    '006083577X',
+    '0385339488',
+    '044024286X',
+    '0385319673',
+    '0312022824',
+    '1250171679',
+    '0307383415',
+    '0316129089',
+    '9781841499918',
+    '9780356504162',
+    '0613035976',
+    '0007137338',
+    '0007137354',
+    '0765350378',
+    '0765350386',
+    '0765356139',
+    '0765316897',
+    '0765330423',
+    '9780765378552',
+    '0765378574',
+    '9780385743563',
+    '0575097418',
+    '0765320304',
+    '0812500482',
+    '0812509714',
+    '0765305119',
+    '0312854315',
+    '0312854277',
+    '0812513754',
+    '0812550285',
+    '0812550293',
+    '0312864256',
+    '0312864590',
+    '0765314673',
+    '0812577566',
+    '0765341530',
+    '0812543971',
+    '0812567609',
+    '0312867018',
+    '0553346687',
+    '0349105324',
+    '0679747044',
+    '0670025291',
+    '067977744X',
+    '0679777458',
+    '0679777466',
+    '0679777474',
+    '0679777482',
+    '0380014300',
+    '0783885040',
+    '0380008319',
+    '0380016648',
+    '0783890648',
+    '0783891601',
+    '0783892942',
+    '0783892918',
+    '0380755017',
+    '0380755025',
+    '9780307477477',
+    '1400043921',
+    '0312140908',
+    '0385721358',
+    '0307387534',
+    '0451461975',
+    '0451458443',
+    '0451458923',
+    '0451459407',
+    '0451461037',
+    '0451461401',
+    '0451461894',
+    '0451462815',
+    '045146317X',
+    '9780356500904',
+    '0452296293',
+    '0765306476',
+    '0441010628',
+    '0441015611',
+    '0808520946',
+    '0575075066',
+    '1857989457',
+    '0425071294',
+    '0399132244',
+    '0765317745',
+    '0765342529',
+    '0425076768',
+    '0765342510',
+    '0450005178',
+    '0425023443',
+    '0765317729',
+    '0765342537',
+    '0765317737',
+    '1591024056',
+    '0553278398',
+    '0553404881',
+    '0553803719',
+    '0380426897',
+    '0785773444',
+    '0553293389',
+    '0586071105',
+    '141041986X',
+    '9780439023498',
+    '9780545310604',
+    '0062024043',
+    '006202406X',
+    '0765313405',
+    '0765313421',
+    '076531343X',
+    '0553386794',
+    '0553381695',
+    '0553381709',
+    '0553801503',
+    '9780007466061',
+    '0765351544',
+    '0312857055',
+    '0312857063',
+    '0312890524',
+    '0312890532',
+    '0812551494',
+    '0312867867',
+    '0765340747',
+    '0765305224',
+    '0765344319',
+    '0765344327',
+    '0765315238',
+    '0765327724',
+    '9780446540339',
+    '1860498876',
+    '0374284385',
+    '0312422156',
+    '144810369X',
+    '1846554071',
+    '0553280589',
+    '0330487965',
+    '015602943X',
+    '0060930187',
+    '0679603069',
+    '0316926345',
+    '9781558496729',
+    '0613709853',
+    '0345331044',
+    '0441005969',
+    '0345324404',
+    '0060273240',
+    '034553980X',
+    '0671004107',
+    '0552159840',
+    '0385721676',
+    '1416556966',
+    '0151001634',
+    '0385534639',
+    '9780007342594',
+    '0061931411',
+    '9780007273812',
+    '0804172447',
+    '0553283685',
+    '0553288202',
+    '0553572946',
+    '055310652X',
+    '0575081953',
+    '0345413369',
+    '0440238153',
+    '0385606990',
+    '9780375845109',
+    '1501133861',
+    '0751512958',
+    '0425200450',
+    '0553380958',
+    '9780380788620',
+    '014011369X',
+    '9780307455772',
+    '0679778314',
+    '1439107955',
+    '0783892780',
+    '0306811081',
+    '0394895894',
+    '0394894537',
+    '0833587560',
+    '0679876154',
+    '0613647882',
+    '0345417984',
+    '034541800X',
+    '0345417968',
+    '034541795X',
+    '0679603352',
+    '0688077080',
+    '0679434968',
+    '0375501371',
+    '0375506276',
+    '1400063833',
+    '9781400063840',
+    '9781451664126',
+    '0756404746'
+  ]
+  end
+end
